+++ flightSoft/trunk/programs/SIPd/SIPd.c
@@ -15,6 +15,7 @@
 #include <signal.h>
 #include <fcntl.h>
 #include <zlib.h>
+#include <libgen.h> //For Mac OS X
 
 /* Flight soft includes */
 #include "sipcomLib/sipcom.h"
@@ -802,8 +803,8 @@ int sendRawWaveformPackets(int bufSize)
     int retVal;
     if(bufSize!=sizeof(AnitaEventBody_t)) {
 	if(errorCounter<100) {
-	    syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(AnitaEventBody_t));
-	    fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(AnitaEventBody_t));
+	  syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(AnitaEventBody_t));
+	  fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(AnitaEventBody_t));
 	    errorCounter++;
 	}
 	return -1;
@@ -838,8 +839,8 @@ int sendRawSurfPackets(int bufSize)
     int retVal;
     if(bufSize!=sizeof(AnitaEventBody_t)) {
 	if(errorCounter<100) {
-	    syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(AnitaEventBody_t));
-	    fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(AnitaEventBody_t));
+	    syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(AnitaEventBody_t));
+	    fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(AnitaEventBody_t));
 	    errorCounter++;
 	}
 	return -1;
@@ -875,8 +876,8 @@ int sendPedSubbedWaveformPackets(int bufSize)
     int retVal;
     if(bufSize!=sizeof(PedSubbedEventBody_t)) {
 	if(errorCounter<100) {
-	    syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(PedSubbedEventBody_t));
-	    fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(PedSubbedEventBody_t));
+	    syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(PedSubbedEventBody_t));
+	    fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(PedSubbedEventBody_t));
 	    errorCounter++;
 	}
 	return -1;
@@ -914,8 +915,8 @@ int sendPedSubbedSurfPackets(int bufSize)
     int retVal;
     if(bufSize!=sizeof(PedSubbedEventBody_t)) {
 	if(errorCounter<100) {
-	    syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(PedSubbedEventBody_t));
-	    fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,sizeof(PedSubbedEventBody_t));
+	    syslog(LOG_ERR,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(PedSubbedEventBody_t));
+	    fprintf(stderr,"Buffer size %d, expected %u -- Bailing\n",bufSize,(unsigned int)sizeof(PedSubbedEventBody_t));
 	    errorCounter++;
 	}
 	return -1;
